Move ConservedVars to fluid module. (#273)
