+++ doc/operators/gas-dynamics.rst
@@ -5,4 +5,5 @@ Gas Dynamics
 .. automodule:: mirgecom.boundary
 .. automodule:: mirgecom.eos
 .. automodule:: mirgecom.initializers
+.. automodule:: mirgecom.inviscid
 .. automodule:: mirgecom.fluid
